+++ tests/growing_inputs_test.cpp
@@ -94,9 +94,9 @@ void GrowingInputsTest::test_perform_inputs_selection(void)
 
     for (size_t i = 0; i < 20; i++)
     {
-        data(i,0) = i;
-        data(i,1) = 10;
-        data(i,2) = i;
+        data(i,0) = (double)i;
+        data(i,1) = 10.0;
+        data(i,2) = (double)i;
     }
 
     ds.set(data);
@@ -125,16 +125,16 @@ void GrowingInputsTest::test_perform_inputs_selection(void)
 
     for (size_t i = 0; i < 10; i++)
     {
-        data(i,0) = j;
-        data(i,1) = 10;
-        data(i,2) = 1;
+        data(i,0) = (double)j;
+        data(i,1) = 10.0;
+        data(i,2) = 1.0;
         j+=1;
     }
     for (size_t i = 10; i < 20; i++)
     {
-        data(i,0) = i;
-        data(i,1) = 10;
-        data(i,2) = 0;
+        data(i,0) = (double)i;
+        data(i,1) = 10.0;
+        data(i,2) = 0.0;
     }
 
     ds.set(data);
